# swagger.yaml
openapi: '3.0.2'
info:
  title: Learn express and swagger - Learn Code Online
  description: LCO - a course segment about writing docs
  version: '1.1.0'
  contact:
    email: allenabraham777@gmail.com
    url: "https://allenkabraham.me"

servers:
  - url: "{protocol}://localhost:3000/api/{version}"
    description: For localhost
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /lco:
    get:
      tags:
        - String
      summary: Returns a greet message from lco 
      responses:
        '200':
          description: All good and success
          content:
            application/json:
              schema:
                type: string
                example: "My String"
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /lcoobject:
    get:
      tags:
        - Objects
      summary: Returns a unique course
      responses:
        '200':
          description: All good and success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
                    example:
                      999
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /courses:
    get:
      tags:
        - "Array"
      summary: Returns all the courses
      responses:
        '200':
          description: All good and success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
                      example:
                        999
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
    post:
      tags:
        - "String"
      summary: Adds a new course to existing courses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
                  example:
                    999
      responses:
        '200':
          description: All good and success
          content:
            application/json:
              schema:
                type: boolean
                example:
                  true
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /courses/{courseId}:
    get:
      tags:
        - "String"
      summary: Returns a course based on request id
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
            default: '22'
      responses:
        '200':
          description: All good and success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
                    example:
                      999
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /courses/query:
    get:
      tags:
        - "String"
      summary: Trying to learn about query
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum:
              - Kochi
              - Bangalore
              - Chennai
            default: Kochi
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum:
              - Web
              - Mobile
            default: Web
      responses:
        '200':
          description: All good and success
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  device:
                    type: string
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /courses/upload:
    post:
      tags:
        - "String"
      summary: Uploading course image
      parameters:
        - in: header
          name: auth
          schema:
            type: string
            default: "Some Header"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: All good and success
          content:
            application/json:
              schema:
                type: boolean
                example: true
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error